test: add tests/core suite (pg_cron app absent, plain db); move core tests
