+++ tests/core/conftest.py
@@ -1,0 +1,6 @@
+from tests.fixtures import (  # noqa: F401
+    _enable_db,
+    _reset_absurd_queues,
+    admin_user,
+    staff_user,
+)
