+++ tests/core/pytest.toml
@@ -1,0 +1,20 @@
+[pytest]
+DJANGO_SETTINGS_MODULE = "tests.core.settings"
+addopts = [
+  "--allow-hosts=db,localhost,127.0.0.1",
+  "--cov",
+  "--cov-report=term",
+  "--cov-report=xml",
+  "--disable-socket",
+  "--junitxml=junit-core.xml",
+  "--reuse-db",
+  "--strict-markers",
+]
+filterwarnings = [
+  "error::bs4.GuessedAtParserWarning",
+]
+markers = [
+  "packaging: builds/inspects the distribution; run only in the dev tox env",
+]
+pythonpath = ["../.."]
+testpaths = ["."]
