+++ tests/core/test_admin_backend_resolve.py
