+++ tests/core/test_admin_views.py
