+++ tests/core/test_app.py
