+++ tests/core/test_queue_sync.py
