+++ tests/core/test_results.py
