+++ tests/core/test_scheduler.py
