+++ tests/core/test_scheduler_checks.py
