+++ tests/settings.py
@@ -9,9 +9,6 @@ INSTALLED_APPS = [
     "django.contrib.sessions",
     "django.contrib.messages",
     "django_absurd",
-    # After "django_absurd": post_migrate receivers fire in INSTALLED_APPS order,
-    # and cron reconcile must run after core queue provisioning.
-    "django_absurd.pg_cron",
     "tests",
 ]
 
@@ -46,7 +43,7 @@ DATABASES = {
         "PASSWORD": os.environ.get("PGPASSWORD", "postgres"),
         "HOST": os.environ.get("PGHOST", "localhost"),
         "PORT": os.environ.get("PGPORT", "5432"),
-        "TEST": {"NAME": "absurd_test"},
+        "TEST": {"NAME": "absurd_test_core"},
     },
     "sqlite": {
         "ENGINE": "django.db.backends.sqlite3",
